requirements.txt
amqp==5.1.1
aniso8601==7.0.0
antlr4-python3-runtime==4.9.3
anyio==3.5.0
appdirs==1.4.3
apturl==0.5.2
argcomplete==1.8.1
argh==0.26.2
asgiref==3.5.2
astor==0.8.1
async-timeout==4.0.2
atomicwrites==1.1.5
attrs==21.4.0
autobahn==22.7.1
Automat==22.10.0
autopep8==1.7.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
billiard==3.6.4.0
black==22.6.0
blinker==1.4
boto3==1.22.6
botocore==1.25.6
Brlapi==0.7.0
build==0.8.0
bytesbufio==1.0.3
cart==1.2.1
celery==5.2.6
certifi==2022.5.18.1
cffi==1.15.0
channels==3.0.4
channels-redis==4.0.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==4.1
django-ajax-selects==2.2.0
django-allauth==0.46.0
django-autoslug==1.9.8
django-ckeditor==6.4.0
django-cors-headers==3.12.0
django-crispy-forms==1.13.0
django-environ==0.8.1
django-etc==1.3.1
django-filter==21.1
django-hitcount==1.3.5
django-js-asset==2.0.0
django-modeltranslation==0.17.3
django-multiselectfield==0.1.12
django-registration-redux==2.9
django-storages==1.12.3
django-taggit==3.0.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.0.1
fastapi==0.87.0
fasteners==0.14.1
filelock==3.0.12
fire==0.4.0
Flask==2.0.2
Flask-API==1.1
fonttools==4.34.4
form==0.0.1
future==0.18.2
gevent==22.10.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.3.post0
gTTS==2.2.4
gunicorn==20.1.0
h11==0.13.0
heroku==0.1.4
html5lib==1.0.1
httplib2==0.14.0
hunspell==0.5.5
hyperlink==21.0.0
idna==3.3
importlib-metadata==1.5.0
incremental==22.10.0
itsdangerous==2.0.1
itypes==1.2.0
jdcal==1.0
Jinja2==3.0.3
jmespath==1.0.0
jsonschema==3.2.0
keyring==18.0.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.4
mypy-extensions==0.4.3
netifaces==0.10.4
numpy==1.23.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
olefile==0.46
omegaconf==2.2.3
opencv-python==4.6.0.66
openpyxl==3.0.3
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.7
paddlepaddle==2.3.2
paramiko==2.6.0
pathspec==0.9.0
pathtools==0.1.2
pbr==5.9.0
pdf2docx==0.5.5
pep517==0.13.0
pexpect==4.6.0
Pillow==9.1.0
pip-tools==6.9.0
platformdirs==2.5.2
pluggy==0.13.0
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.6.1
psycopg==3.0.8
psycopg2-binary==2.9.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.16.2
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.10.1
pycups==1.9.73
pydantic==1.9.0
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==2.4.0
pymacaroons==0.13.0
pympress==1.5.1
PyMuPDF==1.20.1
PyNaCl==1.3.0
pyOpenSSL==22.1.0
pypandoc==1.8.1
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==4.6.9
pytest-runner==6.0.0
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.4
python-docx==0.8.11
python-dotenv==0.20.0
python-pptx==0.6.21
python-vlc==3.0.7110
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
redis==4.3.5
reportlab==3.5.34
requests==2.27.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
Rx==1.6.1
s3transfer==0.5.2
SecretStorage==2.3.1
service-identity==21.1.0
simplejson==3.16.0
SimpleWebSocketServer==0.1.2
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==1.9.5
sqlparse==0.4.2
starlette==0.21.0
systemd-python==234
termcolor==1.1.0
testresources==2.0.1
text-unidecode==1.3
texttable==1.6.4
tk==0.1.0
tkintertable==1.3.3
toml==0.10.2
tomli==2.0.1
torch==1.13.0
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.9
usb-creator==0.3.7
uvicorn==0.17.6
vine==5.0.0
virtualenv==20.0.17
wadllib==1.3.3
watchdog==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
websocket==0.2.1
websocket-client==0.59.0
websockets==8.1
Werkzeug==2.0.2
whitenoise==6.0.0
xkit==0.0.0
XlsxWriter==3.0.3
yapf==0.32.0
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.5.0
